requirements.txt
# Satya Backend Dependencies (Placeholder)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0

# Fault Detection Dependencies
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
attrs==26.1.0
backoff==2.2.1
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
fonttools==4.62.1
frozenlist==1.8.0
idna==3.11
inference-sdk==1.2.0
kiwisolver==1.5.0
marshmallow==3.26.2
matplotlib==3.10.8
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.3.5
opencv-python==4.10.0.84
packaging==26.0
pillow==11.3.0
propcache==0.4.1
py-cpuinfo==9.0.0
pyDeprecate==0.5.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.33.0
scipy==1.17.1
six==1.17.0
supervision==0.27.0.post2
tqdm==4.67.3
typing-inspect==0.9.0
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
